Command gen updates 1.8

+++ src/prompt.h
@@ -1,5 +1,5 @@
 #pragma once
 
-int ask_int(char* text, int lower, int higher);
-long ask_time(char* text);
+long long ask_int(char* text, long long lower, long long higher);
+long long ask_time(char* text);
 int ask_threshold(char* text, int lower, int* voltage_output);
